cleanup bid pr
